// eqrms2/types/group-investor-detail.ts
export type GroupInvestorDetail = {
    investor_id: number;
    created_at: Date;
    group_id: number;
    investor_name: string;
    investor_type: string;
    kyc_status: string;
    kra: string;
    drive_link: string;
    mf_ready: boolean;
    pms_ready: boolean;
    aif_ready: boolean;
    kristal_ready: boolean;
    client_name_pan: string;
    pan_available_bool: boolean;
    pan_number: string;
    poa_india_bool: boolean;
    poa_india_type: string;
    bank_proof_bool: boolean;
    bank_proof_type: string;
    reg_email: string;
    phone_number: string;
    marital_status: string;
    gender: string;
    networth: string;
    annual_income: string;
    wealth_source: string;
    occupation_type: string;
    mother_name: string;
    city_birth: string;
    country_birth: string;
    prime_nominee_name: string;
    prime_nominee_id: string;
    prime_nominee_relation: string;
    poa_overseas_bool: boolean;
    poa_overseas_type: string;
    tax_id: string;
    tax_id_type: string;
    visa_bool: boolean;
    passport_bool: boolean;
    stamp_bool: boolean;
    stamp_date: Date;
    address_details: string;
    other_nominees: string;
    comments_internal: string;
    comments_client_facing: string;
    citizenship_tin_details: string;
};